Add tests to the project, fix bugs

+++ backend/src/main/kotlin/hu/bme/sch/kirpay/account/AccountService.kt
@@ -4,6 +4,7 @@ import hu.bme.sch.kirpay.common.BadRequestException
 import hu.bme.sch.kirpay.common.NotFoundException
 import hu.bme.sch.kirpay.common.RetryTransaction
 import hu.bme.sch.kirpay.principal.getLoggedInPrincipal
+import hu.bme.sch.kirpay.principal.toRef
 import org.springframework.context.ApplicationEventPublisher
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Isolation
@@ -26,35 +27,37 @@ class AccountService(
 
 
   fun create(dto: AccountCreateDto): Account {
-    val account = accountRepository.save(dto.toAccount())
-    events.publishEvent(AccountCreatedEvent(account, getLoggedInPrincipal(), clock.millis()))
+    val account = accountRepository.save(dto.toAccount().copy(id = null))
+    events.publishEvent(AccountCreatedEvent(account, getLoggedInPrincipal()?.toRef(), clock.millis()))
     return account
   }
 
 
   fun setEnabled(accountId: Int, active: Boolean): Account {
     val account = accountRepository.save(find(accountId).copy(active = active))
-    events.publishEvent(AccountUpdatedEvent(account, getLoggedInPrincipal(), clock.millis()))
+    events.publishEvent(AccountUpdatedEvent(account, getLoggedInPrincipal()?.toRef(), clock.millis()))
     return account
   }
 
 
   fun importAccounts(accounts: List<Account>) = accountRepository.saveAll(accounts.map { it.copy(id = null) })
-    .forEach { events.publishEvent(AccountCreatedEvent(it, getLoggedInPrincipal(), clock.millis())) }
+    .forEach { events.publishEvent(AccountCreatedEvent(it, getLoggedInPrincipal()?.toRef(), clock.millis())) }
 
 
   fun update(id: Int, dto: AccountUpdateDto): Account {
     if (!accountRepository.existsById(id)) throw BadRequestException("A számla nem létezik!")
-    val account = accountRepository.save(dto.toAccount(id))
-    events.publishEvent(AccountUpdatedEvent(account, getLoggedInPrincipal(), clock.millis()))
+    val existing = find(id)
+    val account = accountRepository.save(dto.toAccount(id, existing.balance, existing.version))
+    events.publishEvent(AccountUpdatedEvent(account, getLoggedInPrincipal()?.toRef(), clock.millis()))
     return account
   }
 
 
   fun deleteAccount(accountId: Int) {
     val account = find(accountId)
+    if (account.balance > 0) throw BadRequestException("A számla egyenlege nem nulla, nem törölhető!")
     accountRepository.deleteById(accountId)
-    events.publishEvent(AccountDeletedEvent(account, getLoggedInPrincipal(), clock.millis()))
+    events.publishEvent(AccountDeletedEvent(account, getLoggedInPrincipal()?.toRef(), clock.millis()))
   }
 
   fun findActiveByCard(card: String): Account =
@@ -77,11 +80,11 @@ class AccountService(
 
       val updatedHolder = holder.copy(card = null)
       accountRepository.save(updatedHolder)
-      events.publishEvent(AccountUpdatedEvent(updatedHolder, getLoggedInPrincipal(), clock.millis()))
+      events.publishEvent(AccountUpdatedEvent(updatedHolder, getLoggedInPrincipal()?.toRef(), clock.millis()))
     }
 
     val newAccount = accountRepository.save(account.copy(card = card))
-    events.publishEvent(AccountCardAssignedEvent(newAccount, getLoggedInPrincipal(), clock.millis()))
+    events.publishEvent(AccountCardAssignedEvent(newAccount, getLoggedInPrincipal()?.toRef(), clock.millis()))
     return newAccount
   }
